@@ -16,29 +16,27 @@ limitations under the License.
 
 /* joining.js: tests for the various paths when joining a room */
 
-import PlatformPeg from 'matrix-react-sdk/lib/PlatformPeg';
+import PlatformPeg from 'matrix-react-sdk/src/PlatformPeg';
 import WebPlatform from '../../src/vector/platform/WebPlatform';
+import * as sdk from "matrix-react-sdk";
+import * as jssdk from "matrix-js-sdk";
+import "../skin-sdk";
+import "../jest-mocks";
+import React from "react";
+import ReactDOM from "react-dom";
+import ReactTestUtils from "react-dom/test-utils";
+import {makeType} from "matrix-react-sdk/src/utils/TypeUtils";
+import {ValidatedServerConfig} from "matrix-react-sdk/src/utils/AutoDiscoveryUtils";
+import {sleep} from "../test-utils";
+import * as test_utils from "../test-utils";
+import MockHttpBackend from "matrix-mock-request";
+import "fake-indexeddb/auto";
 
-require('skin-sdk');
 
-const jssdk = require('matrix-js-sdk');
-
-const sdk = require('matrix-react-sdk');
 const MatrixChat = sdk.getComponent('structures.MatrixChat');
 const RoomDirectory = sdk.getComponent('structures.RoomDirectory');
 const RoomPreviewBar = sdk.getComponent('rooms.RoomPreviewBar');
 const RoomView = sdk.getComponent('structures.RoomView');
-
-const React = require('react');
-const ReactDOM = require('react-dom');
-const ReactTestUtils = require('react-dom/test-utils');
-const expect = require('expect');
-import {makeType} from "matrix-react-sdk/lib/utils/TypeUtils";
-import {ValidatedServerConfig} from "matrix-react-sdk/lib/utils/AutoDiscoveryUtils";
-import {sleep} from "../test-utils";
-
-const test_utils = require('../test-utils');
-const MockHttpBackend = require('matrix-mock-request');
 
 const HS_URL='http://localhost';
 const IS_URL='http://localhost';
